Fix issue #302: Users cannot be cleansed of decreaseHeal tag or healPrevent tag.
The constant DecreaseHealGivenTag was looking for a literal of "decreaseheal" which was correct.  The issue, is in the list for isNegativeUserEffect, instead of "decreaseheal" it was "decreasehealgiven".  I updated this to "decreaseheal" in the list so that it can be properly found when someone is attempting to cleanse "decreaseheal" effects on them.

I also moved "healprevent" from being in the isPositiveUserEffect list to the isNegativeUserEffect list.  With the healprevent being in the wrong list, if a user effected their opponent with it, it was treated as a positive effect.  So if they were cleared by the enemy, it would fall off instead of staying on.  This was also preventing those effected by it from cleansing themselves of the debuff.

Signed-off-by: Phrosfire <65364337+Phrosfire@users.noreply.github.com>